+++ gui/mainwindow.h
@@ -79,6 +79,7 @@ private:
     void processResponse(TabSession* session, const QJsonArray& apiMessages);
     void handleToolConfirm(TabSession* session, const QJsonObject& toolRequest);
     void handleFinalResponse(TabSession* session, const QJsonObject& response);
+    void handleTurnError(TabSession* session, const QJsonObject& event);
     void handleQuestionRequest(TabSession* session);
     void updateQuickSettingsFor(TabSession* session);